# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Emailibrium API
  version: 0.1.0
  description: |
    Vector-native email intelligence platform API.

    Emailibrium provides semantic search, email classification, ingestion
    orchestration, and inbox intelligence via a REST + SSE interface backed
    by vector embeddings and HNSW indexing.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  contact:
    name: Emailibrium Team
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
...
paths:
...
  /vectors/search/semantic:
    post:
      tags: [vectors]
      summary: Semantic vector search
      description: |
        Embed the query text and perform a cosine similarity search across
        the vector store. Returns ranked results above the minimum score
        threshold.
      operationId: semanticSearch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SemanticSearchRequest"
            example:
              query: "quarterly budget report from finance team"
              limit: 20
              collection: "email_text"
              min_score: 0.5
      responses:
        "200":
          description: Search results ranked by cosine similarity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
              example:
                results:
                  - email_id: "abc-123"
                    score: 0.94
                    collection: "email_text"
                    metadata:
                      category: "Finance"
                  - email_id: "def-456"
                    score: 0.87
                    collection: "email_text"
                    metadata:
                      category: "Work"
                total: 2
                latency_ms: 12
        "500":
          description: Internal server error (embedding or store failure)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /vectors/search/similar/{email_id}:
    post:
      tags: [vectors]
      summary: Find similar emails
      description: |
        Look up the vector for the given email_id and perform a nearest-neighbor
        search to find similar emails. The source email is excluded from results.
      operationId: findSimilar
      parameters:
        - name: email_id
          in: path
          required: true
          schema:
            type: string
          description: The email ID whose similar emails to find
          example: "abc-123"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FindSimilarRequest"
            example:
              limit: 10
      responses:
        "200":
          description: Similar emails ranked by cosine similarity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
        "404":
          description: No vector found for the given email_id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /vectors/classify:
    post:
      tags: [vectors]
      summary: Classify email by category
      description: |
        Generate an embedding for the email content and classify it against
        known category centroids. Uses vector similarity first; falls back to
        LLM classification if confidence is below the configured threshold
        (ADR-004).
      operationId: classifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassifyRequest"
            example:
              email_id: "abc-123"
              subject: "Q4 Budget Review"
              from_addr: "finance@company.com"
              body_text: "Please review the attached Q4 budget spreadsheet."
      responses:
        "200":
          description: Classification result with category and confidence
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassifyResponse"
              example:
                email_id: "abc-123"
                category: "Finance"
                confidence: 0.92
                method: "vector_centroid"
        "500":
          description: Classification failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /vectors/stats:
    get:
      tags: [vectors]
      summary: Vector store statistics
      description: |
        Returns detailed statistics about the vector store including total
        vectors, per-collection counts, dimensionality, memory usage, and
        the index type in use.
      operationId: vectorStats
      responses:
        "200":
          description: Store statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VectorStats"
              example:
                total_vectors: 10523
                collections:
                  email_text: 10000
                  image_text: 523
                dimensions: 384
                memory_bytes: 16834560
                index_type: "in_memory_brute_force"
        "500":
          description: Failed to retrieve statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ---------------------------------------------------------------------------
  # Ingestion
  # ---------------------------------------------------------------------------
...
components:
  schemas:
    SemanticSearchRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
          description: Natural language search query
          example: "quarterly budget report"
        limit:
          type: integer
          description: Maximum number of results (default from server config)
          example: 20
        collection:
          type: string
          enum: [email_text, image_text, image_visual, attachment_text]
          description: Vector collection to search (defaults to email_text)
          example: "email_text"
        filters:
          type: object
          additionalProperties:
            type: string
          description: Metadata key-value filters
          example:
            category: "Finance"
        min_score:
          type: number
          format: float
          description: Minimum cosine similarity score (default from server config)
          example: 0.5
    FindSimilarRequest:
      type: object
      properties:
        limit:
          type: integer
          description: Maximum number of similar emails to return (default 10)
          example: 10
    ClassifyRequest:
      type: object
      required: [email_id, subject, from_addr, body_text]
      properties:
        email_id:
          type: string
          description: Unique identifier of the email to classify
          example: "abc-123"
        subject:
          type: string
          description: Email subject line
          example: "Q4 Budget Review"
        from_addr:
          type: string
          format: email
          description: Sender email address
          example: "finance@company.com"
        body_text:
          type: string
          description: Plain text body of the email
          example: "Please review the attached Q4 budget spreadsheet."
    ClassifyResponse:
      type: object
      properties:
        email_id:
          type: string
          description: The classified email ID
        category:
          type: string
          description: Predicted category
          enum: [Work, Personal, Finance, Shopping, Social, Newsletter, Marketing, Notification, Alerts, Promotions, Uncategorized]
        confidence:
          type: number
          format: float
          description: Confidence score (0.0 to 1.0)
        method:
          type: string
          description: Classification method used
          enum: [vector_centroid, llm_fallback]
    SearchResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/SearchResultItem"
        total:
          type: integer
          description: Number of results returned
        latency_ms:
          type: integer
          format: int64
          description: Server-side search latency in milliseconds
    SearchResultItem:
      type: object
      properties:
        email_id:
          type: string
          description: Source email identifier
        score:
          type: number
          format: float
          description: Cosine similarity score (0.0 to 1.0)
        collection:
          type: string
          description: Vector collection the result belongs to
        metadata:
          type: object
          additionalProperties:
            type: string
          description: Filterable metadata attributes
...
    VectorStats:
      type: object
      properties:
        total_vectors:
          type: integer
          format: int64
          description: Total number of stored vectors
        collections:
          type: object
          additionalProperties:
            type: integer
            format: int64
          description: Vector count per collection
        dimensions:
          type: integer
          description: Embedding dimensionality (e.g., 384 for MiniLM)
        memory_bytes:
          type: integer
          format: int64
          description: Estimated memory usage in bytes
        index_type:
          type: string
          description: Type of vector index in use
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message